Convert all calls to promises
All the calls to Mongo are now using promises instead of callbacks.

Also there is a general error handler for error.
This handler responds with 500 every time there's an error,
and logs the error on the server side without exposing it to the user.

Adding a fixer function for fixing values to functions :)